Refactor code,undo State Engine State Use,Now On Standby
Refactored code in Megaman script,enemy_template script and
working on GlobalScript code to be snake_case instead of
camelCase for easy reading and editing

Undid the use of the state engine by disabling the initial movement and attack states.
This is to prevent confusion + movement using the state engine needs to be worked on to feel good.